update document interface

+++ apps/document_interface/services.py
@@ -15,6 +15,8 @@ from django.db import transaction
 from asgiref.sync import sync_to_async
 
 from .models import DocumentMessage, DocumentChunk, DocumentMetadata
+from .helper import generate_pdf_hash
+
 
 API_URL = "https://api-inference.huggingface.co/models/facebook/bart-large-cnn"
 headers = {"Authorization": f"Bearer {settings.HUGGINGFACE_API_TOKEN}"}
@@ -30,47 +32,54 @@ class DocumentModalHandler:
         # Decode the base64 content
         pdf_content = base64.b64decode(document_content)
 
-        # Create a temporary file
-        with tempfile.NamedTemporaryFile(delete=False, suffix='.pdf', mode='wb') as temp_file:
-            temp_file.write(pdf_content)
-            temp_file_path = temp_file.name
+        # Generate hash for the pdf document
+        document_hash = generate_pdf_hash(pdf_content)
 
-        try:
-            # Load the PDF using the temporary file path
-            loader = PyPDFLoader(temp_file_path)
-            pages = loader.load()
+        # Check if the document already exists
+        existing_document = await self.get_existing_document(document_hash)
+        if existing_document:
+            return await self.get_existing_document_info(existing_document)
 
-            # Extract metadata
-            metadata = await self.extract_metadata(temp_file_path)
+        else:
+            # Create a temporary file
+            with tempfile.NamedTemporaryFile(delete=False, suffix='.pdf', mode='wb') as temp_file:
+                temp_file.write(pdf_content)
+                temp_file_path = temp_file.name
 
-            # Split the document into chunks
-            text_splitter = RecursiveCharacterTextSplitter(
-                chunk_size=1000,
-                chunk_overlap=200,
-                length_function=len,
-            )
-            chunks = text_splitter.split_documents(pages)
+            try:
+                # Load the PDF using the temporary file path
+                loader = PyPDFLoader(temp_file_path)
+                pages = loader.load()
 
-            # Generate embeddings for chunks
-            embeddings = await asyncio.to_thread(self.embeddings.embed_documents, [chunk.page_content for chunk in chunks])
+                # Extract metadata
+                metadata = await self.extract_metadata(temp_file_path)
 
-            # Generate summary
-            full_text = " ".join([chunk.page_content for chunk in chunks])
-            summary = await self.generate_summary_using_LLM(full_text[:4000])
+                # Split the document into chunks
+                text_splitter = RecursiveCharacterTextSplitter(
+                    chunk_size=1000,
+                    chunk_overlap=200,
+                    length_function=len,
+                )
+                chunks = text_splitter.split_documents(pages)
 
-            return len(chunks), summary, chunks, embeddings, metadata
-        finally:
-            # clean up the temporary file
-            os.unlink(temp_file_path)
+                # Generate embeddings for chunks
+                embeddings = await asyncio.to_thread(self.embeddings.embed_documents, [chunk.page_content for chunk in chunks])
+
+                # Generate summary
+                full_text = " ".join([chunk.page_content for chunk in chunks])
+                summary = await self.generate_summary_using_LLM(full_text[:4000])
+
+                return len(chunks), summary, chunks, embeddings, metadata, document_hash
+            finally:
+                # clean up the temporary file
+                os.unlink(temp_file_path)
 
     async def extract_metadata(self, file):
         pdf = pypdf.PdfReader(file)
         info = pdf.metadata
-        print(info)
         return {
             'title': info.get('/Title', ''),
             'author': info.get('/Author', ''),
-            'creation_date': info.get('/CreationDate', ''),
             'page_count': len(pdf.pages),
             'word_count': sum(len(page.extract_text().split()) for page in pdf.pages)
         }
@@ -94,13 +103,12 @@ class DocumentModalHandler:
                 key=lambda x: self.cosine_similarity(
                     query_embedding, x.embedding),
                 reverse=True
-            )[:3]
+            )[:2]
             return relevant_chunks
 
         # Embed the query and retirve related chunks
         query_embedding = await asyncio.to_thread(self.embeddings.embed_query, query)
         relevant_chunks = await retrive_relevant_chunk(query_embedding)
-
         context = " ".join([chunk.content for chunk in relevant_chunks])
         return context
 
@@ -108,8 +116,31 @@ class DocumentModalHandler:
     def cosine_similarity(v1, v2):
         return sum(a*b for a, b in zip(v1, v2)) / (sum(a*a for a in v1)**0.5 * sum(b*b for b in v2)**0.5)
 
+    @sync_to_async
+    def get_existing_document(self, document_hash):
+        try:
+            return DocumentMessage.objects.get(document_hash=document_hash)
+        except DocumentMessage.DoesNotExist:
+            return None
+
+    @sync_to_async
+    def get_existing_document_info(self, document):
+        return (
+            document.num_chunks,
+            document.summary,
+            None,  # We don't need to return chunks for an existing document
+            None,  # We don't need to return embeddings for an existing document
+            {
+                'title': document.metadata.title,
+                'author': document.metadata.author,
+                'page_count': document.metadata.page_count,
+                'word_count': document.metadata.word_count
+            },
+            document.document_hash
+        )
+
     @staticmethod
-    async def save_document_message(message, document_data, num_chunks, summary, chunks, embeddings, metadata):
+    async def save_document_message(message, document_data, num_chunks, summary, chunks, embeddings, metadata, document_hash):
         @sync_to_async
         def create_document_message():
             document_content = ContentFile(base64.b64decode(
@@ -119,22 +150,24 @@ class DocumentModalHandler:
                 document=document_content,
                 num_pages=metadata['page_count'],
                 num_chunks=num_chunks,
-                processed_content=" ".join(
+                processed_content="" if chunks is None else " ".join(
                     [chunk.page_content for chunk in chunks]),
-                summary=summary
+                summary=summary,
+                document_hash=document_hash
             )
 
         @sync_to_async
         def create_chunks_and_metadata(document_message):
             with transaction.atomic():
                 # Save chunks and embeddings
-                for chunk, embedding in zip(chunks, embeddings):
-                    DocumentChunk.objects.create(
-                        document=document_message,
-                        content=chunk.page_content,
-                        embedding=embedding,
-                        metadata=chunk.metadata
-                    )
+                if chunks and embeddings:
+                    for chunk, embedding in zip(chunks, embeddings):
+                        DocumentChunk.objects.create(
+                            document=document_message,
+                            content=chunk.page_content,
+                            embedding=embedding,
+                            metadata=chunk.metadata
+                        )
 
                 # Save metadata
                 DocumentMetadata.objects.create(
